added script and new generated text in script folder
