--- script/checkpoint/run1/hparams.json
@@ -1,7 +1,0 @@
-{
-  "n_vocab": 50257,
-  "n_ctx": 1024,
-  "n_embd": 1024,
-  "n_head": 16,
-  "n_layer": 24
-}
