--- script/requirements.txt
@@ -1,40 +1,0 @@
-absl-py==0.8.0
-astor==0.8.0
-blis==0.2.4
-certifi==2019.9.11
-chardet==3.0.4
-cymem==2.0.2
-en-core-web-sm==2.1.0
-gast==0.3.2
-google-pasta==0.1.7
-gpt-2-simple==0.6
-grpcio==1.23.0
-h5py==2.10.0
-idna==2.8
-Keras-Applications==1.0.8
-Keras-Preprocessing==1.1.0
-Markdown==3.1.1
-murmurhash==1.0.2
-numpy==1.17.2
-pandas==0.25.1
-plac==0.9.6
-preshed==2.0.1
-protobuf==3.9.1
-python-dateutil==2.8.0
-pytz==2019.2
-regex==2019.8.19
-requests==2.22.0
-six==1.12.0
-spacy==2.1.8
-srsly==0.1.0
-tensorboard==1.14.0
-tensorflow==1.14.0
-tensorflow-estimator==1.14.0
-termcolor==1.1.0
-thinc==7.0.8
-toposort==1.5
-tqdm==4.35.0
-urllib3==1.25.3
-wasabi==0.2.2
-Werkzeug==0.15.6
-wrapt==1.11.2
